+++ packages/hono-takibi/src/generator/zod-openapi-hono/handler/helper/group-handlers-by-file-name.ts
@@ -1,4 +1,4 @@
-export function groupHandlersByFileNameHelper(
+export function groupHandlersByFileName(
   handlers: {
     fileName: `${string}.ts`
     testFileName: `${string}.ts`
